requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
tensorflow==1.12.0